reorganized __test__ folder structure // all tests pass
