+++ SmartClothingApp/__test__/__unit-test__/__actions-test__/userActions.unit.test.js
@@ -1,10 +1,10 @@
-import { auth, database } from '../firebaseConfig.js';
+import { auth, database } from '../../../firebaseConfig.js';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { firebaseErrorsMessages } from '../src/utils/firebaseErrorsMessages.js';
+import { firebaseErrorsMessages } from '../../../src/utils/firebaseErrorsMessages.js';
 import { render, waitFor } from "@testing-library/react"
 import flushPromises from 'flush-promises';
-import { storeUID, storeMetrics } from "../src/utils/localStorage.js";
+import { storeUID, storeMetrics } from "../../../src/utils/localStorage.js";
 import { 
   setDoc, 
   getDoc, 
@@ -33,7 +33,7 @@ import {
   queryHeartRateData,
   deleteAccount,
   logout,
-} from '../src/actions/userActions.js'; 
+} from '../../../src/actions/userActions.js'; 
 import { 
   LOGIN_WITH_EMAIL,
   SIGNUP_WITH_EMAIL,
@@ -42,7 +42,7 @@ import {
   UPDATE_USER_METRICS_DATA,
   UPDATE_EMAIL_SUCCESS,
   UPDATE_PASSWORD_SUCCESS,
-} from '../src/actions/types';
+} from '../../../src/actions/types.js';
 import {
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
@@ -54,14 +54,14 @@ import {
   updatePassword,
   signOut,
 } from 'firebase/auth';
-import { toastError } from "../src/actions/toastActions.js";
+import { toastError } from "../../../src/actions/toastActions.js";
 import { 
   userMetricsDataModalVisible 
-} from '../src/actions/appActions';
+} from '../../../src/actions/appActions.js';
 import { type } from '@testing-library/react-native/build/user-event/type/type.js';
 
 
-jest.mock('../src/utils/localStorage.js', () => ({
+jest.mock('../../../src/utils/localStorage.js', () => ({
   AsyncStorage: jest.fn(),
   storeUID: jest.fn(),
   storeMetrics: jest.fn()
@@ -82,7 +82,7 @@ jest.mock('firebase/firestore', () => ({
   deleteDoc: jest.fn(),
 }));
 
-jest.mock('../firebaseConfig.js', () => ({
+jest.mock('../../../firebaseConfig.js', () => ({
   auth: {
     signOut: jest.fn(() => Promise.resolve()),
     currentUser: {
@@ -114,14 +114,14 @@ jest.mock('firebase/auth', () => ({
   },
 }));
 
-jest.mock('../src/actions/appActions', () => ({
+jest.mock('../../../src/actions/appActions', () => ({
   userMetricsDataModalVisible: jest.fn(() => ({
     type: 'USER_METRICS_DATA_MODAL_VISIBLE',
     payload: { visibility: true, isFromSignUpScreen: true },
   })),
 }));
 
-jest.mock('../src/actions/toastActions', () => ({
+jest.mock('../../../src/actions/toastActions', () => ({
   toastError: jest.fn((message) => ({
     type: 'showErrorToast',
     payload: message,
